Add plaintext plugin name to discord hook payload
When replying to a feedback message (not uncommon, I've seen) people pretty much have to click the jump-to-reply-target link to see what feedback is being talked about. This would just add a single normal-text line with the plugin name (which the report window takes automatically, so users can't just post anything in there) so that the reply will show the plugin in question.