Silently drop feedback with refused contact info
Any feedback with contact info containing the words "just" and "feedback" or that consists solely of the word "feedback" will be dropped, with no indication to the user that it has happened so as to prevent them from attempting to circumvent this filter. The list of filters is easily extensible and has been documented.

This will stop the feedback channel from having to put up with that one user who keeps sending aggressive all-caps copy-pasted paragraphs of a single sentence. That is not a strawman, that exact user exists and has sent numerous such feedbacks.